Move `Commander` and `StatCommander` to own package.
