+++ cmd/cmd.go
@@ -1,0 +1,41 @@
+// Copyright 2016 Mender Software AS
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//        http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+package cmd
+
+import (
+	"os"
+	"os/exec"
+)
+
+type Commander interface {
+	Command(name string, arg ...string) *exec.Cmd
+}
+
+type StatCommander interface {
+	Stat(string) (os.FileInfo, error)
+	Commander
+}
+
+// we need real OS implementation
+type OsCalls struct {
+}
+
+func (OsCalls) Command(name string, arg ...string) *exec.Cmd {
+	return exec.Command(name, arg...)
+}
+
+func (OsCalls) Stat(name string) (os.FileInfo, error) {
+	return os.Stat(name)
+}
